==> 01_Preprocess.Rmd <==
---
title: "01. Preprocess"
subtitle: "December 15th, 2023"
author: "Ryohei SHIBATA"
---

```{R, include = FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = FALSE, 
                      eval = FALSE, 
                      warning = FALSE, 
                      message = FALSE, 
                      comment = "",
                      cache = FALSE) 

Sys.setlocale("LC_ALL",
              'UTF-8')
options(scipen = 10)
```

```{R}
source("Library.R")
source("Function.R")
source("FeatureSelection_Dualomics.R")
source("FeatureSelection_Clinical.R")
```

# Note
## Coding rules
* `_tbw`, tibble class object (wide type)
* `_tbl`, tibble class object (long type, i.e., tidy data)
* `_df`, data frame class object

## Objects
* `Clinicaldata_tbw`, clinical data
* `mRNA_RA_RAW_tbl`,  transcriptome data
* `Lipid_Species_tbl`, lipidome data

# Clinicaldata
## Determine study_id for analysis
```{R}
Clinicaldata_tbw %>% 
  select(study_id, asthma_epi_6yr) %>% 
       filter(!is.na(asthma_epi_6yr)) -> Asthma_tbw

intersect(Asthma_tbw %>% 
            check_levels("study_id"),
          mRNA_RA_RAW_tbl %>% 
            check_levels("study_id")) %>% 
  intersect(., 
            Lipid_Species_tbl %>% 
            check_levels("study_id")) %>% 
  as.character -> study_id_Analysis # 342
```

## Select clinical features
```{R}
Clinicaldata_tbw %>% 
  mutate_if(is.factor, droplevels) %>% 
  c2f -> Clinicaldata_LR_tbw

Clinicaldata_LR_tbw %>% 
  select_if(is.factor) %>% 
  gather(Variable, Value, -study_id) %>% 
  group_by(Variable, Value) %>% 
  summarise() %>% 
  group_by(Variable) %>% 
  summarise(Count = n()) %>% 
  filter(Count == 1) %>% 
  .$Variable -> Variable_Onelevel

Clinicaldata_LR_tbw %>% 
  select(!Variable_Onelevel) %>% 
  select(-Age_mo, -intake_sex, -raceethn) %>% 
  names() %>% 
  .[-1] -> Exposures

map(Exposures, function(x){
  Clinicaldata_LR_tbw %>% 
        inner_join(Asthma_tbw) %>% 
    rename(Exposure = x) %>% 
  glm(asthma_epi_6yr ~ Exposure + Age_mo + intake_sex + raceethn,
      family = "binomial"(link = "logit"),
      data = .) %>% 
      broom::tidy(.,
                  conf.int = TRUE,
                  exponentiate = TRUE) %>% 
    mutate(Variable = x)})  %>% 
  do.call(bind_rows, .) %>% 
  filter(grepl("Exposure", term)) %>% 
  filter(p.value < 0.05) %>% 
  mutate(term = Variable) %>% 
  select(-Variable) %>% 
  arrange(p.value) %>% 
  mutate(P.value = as.character(p.value)) %>% 
  tidy_table() %>% 
  mutate(p.value = as.numeric(P.value)) %>% 
  select(-P.value) %>% 
  rename(Variable = term) -> LR_Clinical_tbw

LR_Clinical_tbw %>% 
  .$Variable -> Feature_Clinical
```

```{R}
Clinicaldata_LR_tbw %>%
  filter(study_id %in% study_id_Analysis) %>%
  select(study_id,
         Feature_Clinical) -> Clinicaldata_NN_tbw

Clinicaldata_NN_tbw %>% 
  mutate(RVC_NPAseq = if_else(is.na(RVC_NPAseq), 0, 1)) %>%
  mutate(prev_breathingprob = if_else(prev_breathingprob == "0", 0, 1)) %>% 
  mutate_if(is.factor, as.character) %>% 
  mutate_if(is.character, as.numeric) %>% 
  gather(Variable, Value, -study_id) %>%
  mutate(Value = as.numeric(as.character(Value))) %>%
  mutate(study_id = as.factor(study_id)) %>% 
  c2f -> Clinical_tbl

Clinicaldata_NN_tbw %>% 
  select_if(is.numeric) %>% 
  names -> Variable_Con
```

## Select omics features and generate interaction variables
```{R}
list(Omics1 = mRNA_RA_RAW_tbl %>% 
       ungroup %>% 
       rename(Variable = Ensembl) %>% 
       filter(study_id %in% study_id_Analysis) %>% 
       remove_all0 %>% 
       remove_iqr0 %>% 
       c2f,
     
     Omics2 = Lipid_Species_tbl %>% 
       filter(study_id %in% study_id_Analysis) %>%
       ungroup %>% 
      remove_all0 %>% 
       remove_iqr0 %>% 
       c2f
     ) -> Dualomics_tbl_list

# Model
Feature_model <- as.formula("asthma_epi_6yr ~ Value + 
                                   raceethn + 
                                   Age_mo + intake_sex")
```

```{R}
Pvalue_Cutoff1 <- 0.05
Pvalue_Cutoff2 <- 0.00005
Pvalue_Cutoff3 <- 0.00005

source("FeatureSelection_Dualomics_YJT.R")
source("FeatureSelection_Clinical11.R")
```

# Output
## Generate .csv to transfer data into Python
```{R}
ClinicalDualomicsProduct_tbw %>% 
  arrange(study_id) %>% 
  c2r("study_id") %>% 
  .[study_id_Analysis, , drop = FALSE] -> x_df

write_csv(x_df,
          "ClinicalmRNALipidProduct_Value.csv")
  
Asthma_tbw %>% 
  filter(study_id %in% study_id_Analysis) %>% 
  arrange(study_id) %>% 
  c2r("study_id") %>% 
  .[study_id_Analysis, , drop = FALSE] -> y_df

write_csv(y_df,
          "ClinicalmRNALipidProduct_Asthma.csv")
```
